requirements.txt
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
sentencepiece==0.1.96
numpy==1.18.5
tensorboardX==1.8
nltk==3.4.5
pyrouge==0.1.3
rouge==0.3.2
spacy==3.0.6
html2text==2020.1.16
transformers==4.6.1
python-telegram-bot==13.5
requests==2.25.1
rasa==2.6.2
python-dotenv==0.17.1